[update] Init update Db . Update dao Roadmap

+++ data/src/main/java/com/gkreduction/data/entity/SubtopicRemote.kt
@@ -5,6 +5,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 class SubtopicRemote(
-    @SerialName("id") val id: Int?,
+    @SerialName("id") val id: Long?,
     @SerialName("name") val name: String?,
 ) : BaseRemote
