+++ domain/src/main/java/com/gkreduction/domain/entity/BaseElement.kt
@@ -1,6 +1,6 @@
 package com.gkreduction.domain.entity
 
-abstract class BaseElement(open val id: Int, open var name: String) {
+abstract class BaseElement(open val id: Long, open var name: String) {
     override fun equals(other: Any?): Boolean {
         if (other is BaseElement)
             return name == other.name
@@ -8,6 +8,6 @@ abstract class BaseElement(open val id: Int, open var name: String) {
     }
 
     override fun hashCode(): Int {
-        return 31 * id + name.hashCode()
+        return (31 * id + name.hashCode()).toInt()
     }
 }
