+++ domain/src/main/java/com/gkreduction/domain/entity/Subtopic.kt
@@ -1,3 +1,3 @@
 package com.gkreduction.domain.entity
 
-class Subtopic(override var id: Int,override var name: String) : BaseElement(id, name)
+class Subtopic(override var id: Long,override var name: String) : BaseElement(id, name)
