+++ domain/src/main/java/com/gkreduction/domain/entity/Topic.kt
@@ -1,4 +1,4 @@
 package com.gkreduction.domain.entity
 
-class Topic(override var id: Int, override var name: String, var subtopics: List<BaseElement>) :
+class Topic(override var id: Long, override var name: String, var subtopics: List<BaseElement>) :
     BaseElement(id, name)
